requirements.txt
streamlit
google-genai
beautifulsoup4
playwright
python-docx
reportlab
requests
python-dotenv
pandas
